move tools-bar
