+++ src/app/pages/tools-bar/tools-bar.component.scss
